refactor(init-config): setSetting as fire-and-forget (drop overlay)
Research across Neovim, VS Code, Emacs, Zed, and Sublime shows no
editor tracks per-plugin setting writes. All use fire-and-forget into
shared state. Align with the ecosystem.

Deleted:
  - runtime_config.rs (RuntimeConfigOverlay, set_dot_path tests)
  - base_config_json field on Editor
  - runtime_overlay field on Editor
  - rebuild_config_from_overlay method
  - clear_runtime_overlay method
  - ClearRuntimeOverlay PluginCommand variant (already gone)
  - Overlay init in Editor construction
  - Overlay clear in load_init_script and InitRevert

Added:
  - set_dot_path as a private fn in editor_init.rs (the one useful
    bit from the deleted module)

Changed:
  - handle_set_setting: serializes current config to JSON, patches
    the path, deserializes back, assigns via Arc::new. Propagates
    theme/keybinding/bar-visibility side effects. On validation
    failure, surfaces an error and keeps the prior config.
  - reload_config (toggle_actions.rs): removed overlay replay — disk
    config is the ground truth, setSetting writes from any plugin
    persist until overwritten or editor restart.
  - InitRevert: unloads the plugin (commands, handlers, events,
    exported APIs) but setSetting writes persist. Consistent with
    how registerLspServer/applyTheme already work.
  - Tests: reloading_init_ts_reverts_prior_setting renamed to
    set_setting_is_fire_and_forget_across_reload (asserts the
    value persists). init_revert_action_drops_prior_setting_writes
    renamed to init_revert_unloads_plugin_but_settings_persist
    (asserts settings survive revert).

Net: ~200 lines deleted, setSetting is now consistent with every
other mutating plugin API call in the editor.
