+++ crates/fresh-editor/src/lib.rs
@@ -23,8 +23,6 @@ pub mod config_io;
 #[cfg(feature = "runtime")]
 pub mod init_script;
 #[cfg(feature = "runtime")]
-pub mod runtime_config;
-#[cfg(feature = "runtime")]
 pub mod state;
 #[cfg(feature = "runtime")]
 pub mod workspace;
